Build the seeder for listing_tag pivot table
From User to Listing and from Listing to Tag, through this chain, build
the listing_tag pivot table.  Both Listing and Tag are associated with
User.

Add timestamps to listing_tag pivot table as Listing and Tag work
together to seed the pivot table.

Add the seeder for the tags table to the root seeder.